Reorganized layers
