+++ machine_comprehension/models/layers/answer_point_layers.py
@@ -1,0 +1,230 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+__author__ = 'han'
+
+import h5py
+import torch
+import torch.nn.functional as F
+import numpy as np
+from ...utils.functions import masked_softmax, compute_mask, masked_flip
+
+
+class PointerAttention(torch.nn.Module):
+    r"""
+    attention mechanism in pointer network
+    Args:
+        - input_size: The number of features in Hr
+        - hidden_size: The number of features in the hidden layer
+
+    Inputs:
+        Hr(context_len, batch, hidden_size * num_directions): question-aware context representation
+        Hk_last(batch, hidden_size): the last hidden output of previous time
+
+    Outputs:
+        beta(batch, context_len): question-aware context representation
+    """
+
+    def __init__(self, input_size, hidden_size):
+        super(PointerAttention, self).__init__()
+
+        self.linear_wr = torch.nn.Linear(input_size, hidden_size)
+        self.linear_wa = torch.nn.Linear(hidden_size, hidden_size)
+        self.linear_wf = torch.nn.Linear(hidden_size, 1)
+
+    def forward(self, Hr, Hr_mask, Hk_pre):
+        wr_hr = self.linear_wr(Hr)  # (context_len, batch, hidden_size)
+        wa_ha = self.linear_wa(Hk_pre).unsqueeze(0)  # (1, batch, hidden_size)
+        f = F.tanh(wr_hr + wa_ha)  # (context_len, batch, hidden_size)
+
+        beta_tmp = self.linear_wf(f) \
+            .squeeze(2) \
+            .transpose(0, 1)  # (batch, context_len)
+
+        beta = masked_softmax(beta_tmp, m=Hr_mask, dim=1)
+        return beta
+
+
+class SeqPointer(torch.nn.Module):
+    r"""
+    Sequence Pointer Net that output every possible answer position in context
+    Args:
+
+    Inputs:
+        Hr: question-aware context representation
+    Outputs:
+        **output** every answer index possibility position in context, no fixed length
+    """
+
+    def __init__(self):
+        super(SeqPointer, self).__init__()
+
+    def forward(self, *input):
+        return NotImplementedError
+
+
+class UniBoundaryPointer(torch.nn.Module):
+    r"""
+    Unidirectional Boundary Pointer Net that output start and end possible answer position in context
+    Args:
+        - input_size: The number of features in Hr
+        - hidden_size: The number of features in the hidden layer
+        - enable_layer_norm: Whether use layer normalization
+
+    Inputs:
+        Hr(context_len, batch, hidden_size * num_directions): question-aware context representation
+        h_0(batch, hidden_size): init lstm cell hidden state
+    Outputs:
+        **output** (answer_len, batch, context_len): start and end answer index possibility position in context
+        **hidden** (batch, hidden_size), [(batch, hidden_size)]: rnn last state
+    """
+    answer_len = 2
+
+    def __init__(self, mode, input_size, hidden_size, enable_layer_norm):
+        super(UniBoundaryPointer, self).__init__()
+
+        self.input_size = input_size
+        self.hidden_size = hidden_size
+        self.enable_layer_norm = enable_layer_norm
+
+        self.attention = PointerAttention(input_size, hidden_size)
+
+        if self.enable_layer_norm:
+            self.layer_norm = torch.nn.LayerNorm(input_size)
+
+        self.mode = mode
+        if mode == 'LSTM':
+            self.hidden_cell = torch.nn.LSTMCell(input_size, hidden_size)
+        elif mode == 'GRU':
+            self.hidden_cell = torch.nn.GRUCell(input_size, hidden_size)
+        else:
+            raise ValueError('Wrong mode select %s, change to LSTM or GRU' % mode)
+        self.reset_parameters()
+
+    def reset_parameters(self):
+        """
+        Here we reproduce Keras default initialization weights to initialize Embeddings/LSTM weights
+        """
+        ih = (param for name, param in self.named_parameters() if 'weight_ih' in name)
+        hh = (param for name, param in self.named_parameters() if 'weight_hh' in name)
+        b = (param for name, param in self.named_parameters() if 'bias' in name)
+        for t in ih:
+            torch.nn.init.xavier_uniform_(t)
+        for t in hh:
+            torch.nn.init.orthogonal_(t)
+        for t in b:
+            torch.nn.init.constant_(t, 0)
+
+    def forward(self, Hr, Hr_mask, h_0=None):
+        if h_0 is None:
+            batch_size = Hr.shape[1]
+            h_0 = Hr.new_zeros(batch_size, self.hidden_size)
+
+        hidden = (h_0, h_0) if self.mode == 'LSTM' and isinstance(h_0, torch.Tensor) else h_0
+        beta_out = []
+
+        for t in range(self.answer_len):
+            attention_input = hidden[0] if self.mode == 'LSTM' else hidden
+            beta = self.attention.forward(Hr, Hr_mask, attention_input)  # (batch, context_len)
+            beta_out.append(beta)
+
+            context_beta = torch.bmm(beta.unsqueeze(1), Hr.transpose(0, 1)) \
+                .squeeze(1)  # (batch, input_size)
+
+            if self.enable_layer_norm:
+                context_beta = self.layer_norm(context_beta)  # (batch, input_size)
+
+            hidden = self.hidden_cell.forward(context_beta, hidden)  # (batch, hidden_size), (batch, hidden_size)
+
+        result = torch.stack(beta_out, dim=0)
+        return result, hidden
+
+
+class BoundaryPointer(torch.nn.Module):
+    r"""
+    Boundary Pointer Net that output start and end possible answer position in context
+    Args:
+        - input_size: The number of features in Hr
+        - hidden_size: The number of features in the hidden layer
+        - bidirectional: Bidirectional or Unidirectional
+        - dropout_p: Dropout probability
+        - enable_layer_norm: Whether use layer normalization
+
+    Inputs:
+        Hr (context_len, batch, hidden_size * num_directions): question-aware context representation
+        h_0 (batch, hidden_size): init lstm cell hidden state
+    Outputs:
+        **output** (answer_len, batch, context_len): start and end answer index possibility position in context
+    """
+
+    def __init__(self, mode, input_size, hidden_size, bidirectional, dropout_p, enable_layer_norm):
+        super(BoundaryPointer, self).__init__()
+        self.bidirectional = bidirectional
+        self.hidden_size = hidden_size
+
+        self.left_ptr_rnn = UniBoundaryPointer(mode, input_size, hidden_size, enable_layer_norm)
+        if bidirectional:
+            self.right_ptr_rnn = UniBoundaryPointer(mode, input_size, hidden_size, enable_layer_norm)
+
+        self.dropout = torch.nn.Dropout(p=dropout_p)
+
+    def forward(self, Hr, Hr_mask, h_0=None):
+        Hr = self.dropout.forward(Hr)
+
+        left_beta, _ = self.left_ptr_rnn.forward(Hr, Hr_mask, h_0)
+        rtn_beta = left_beta
+        if self.bidirectional:
+            right_beta_inv, _ = self.right_ptr_rnn.forward(Hr, Hr_mask, h_0)
+            right_beta = right_beta_inv[[1, 0], :]
+
+            rtn_beta = (left_beta + right_beta) / 2
+
+        # todo: unexplainable
+        new_mask = torch.neg((Hr_mask - 1) * 1e-6)  # mask replace zeros with 1e-6, make sure no gradient explosion
+        rtn_beta = rtn_beta + new_mask.unsqueeze(0)
+
+        return rtn_beta
+
+
+class MultiHopBdPointer(torch.nn.Module):
+    r"""
+    Boundary Pointer Net with Multi-Hops that output start and end possible answer position in context
+    Args:
+        - input_size: The number of features in Hr
+        - hidden_size: The number of features in the hidden layer
+        - num_hops: Number of max hops
+        - dropout_p: Dropout probability
+        - enable_layer_norm: Whether use layer normalization
+
+    Inputs:
+        Hr (context_len, batch, hidden_size * num_directions): question-aware context representation
+        h_0 (batch, hidden_size): init lstm cell hidden state
+    Outputs:
+        **output** (answer_len, batch, context_len): start and end answer index possibility position in context
+    """
+
+    def __init__(self, mode, input_size, hidden_size, num_hops, dropout_p, enable_layer_norm):
+        super(MultiHopBdPointer, self).__init__()
+        self.hidden_size = hidden_size
+        self.num_hops = num_hops
+
+        self.ptr_rnn = UniBoundaryPointer(mode, input_size, hidden_size, enable_layer_norm)
+        self.dropout = torch.nn.Dropout(p=dropout_p)
+
+    def forward(self, Hr, Hr_mask, h_0=None):
+        Hr = self.dropout.forward(Hr)
+
+        beta_last = None
+        for i in range(self.num_hops):
+            beta, h_0 = self.ptr_rnn.forward(Hr, Hr_mask, h_0)
+            if beta_last is not None and (beta_last == beta).sum().item() == beta.shape[0]:  # beta not changed
+                break
+
+            beta_last = beta
+
+        new_mask = torch.neg((Hr_mask - 1) * 1e-6)  # mask replace zeros with 1e-6, make sure no gradient explosion
+        rtn_beta = beta + new_mask.unsqueeze(0)
+
+        return rtn_beta
+
+
