+++ machine_comprehension/models/layers/match_layers.py
@@ -1,0 +1,197 @@
+import h5py
+import torch
+import torch.nn.functional as F
+import numpy as np
+from ...utils.functions import masked_softmax, compute_mask, masked_flip
+
+
+
+class MatchRNNAttention(torch.nn.Module):
+    r"""
+    attention mechanism in match-rnn
+    Args:
+        - input_size: The number of expected features in the input Hp and Hq
+        - hidden_size: The number of features in the hidden state Hr
+
+    Inputs:
+        Hpi(batch, input_size): a context word encoded
+        Hq(question_len, batch, input_size): whole question encoded
+        Hr_last(batch, hidden_size): last lstm hidden output
+
+    Outputs:
+        alpha(batch, question_len): attention vector
+    """
+
+    def __init__(self, hp_input_size, hq_input_size, hidden_size):
+        super(MatchRNNAttention, self).__init__()
+
+        self.linear_wq = torch.nn.Linear(hq_input_size, hidden_size)
+        self.linear_wp = torch.nn.Linear(hp_input_size, hidden_size)
+        self.linear_wr = torch.nn.Linear(hidden_size, hidden_size)
+        self.linear_wg = torch.nn.Linear(hidden_size, 1)
+
+    def forward(self, Hpi, Hq, Hr_last):
+        wq_hq = self.linear_wq(Hq)  # (question_len, batch, hidden_size)
+        wp_hp = self.linear_wp(Hpi).unsqueeze(0)  # (1, batch, hidden_size)
+        wr_hr = self.linear_wr(Hr_last).unsqueeze(0)  # (1, batch, hidden_size)
+        G = F.tanh(wq_hq + wp_hp + wr_hr)  # (question_len, batch, hidden_size), auto broadcast
+        wg_g = self.linear_wg(G) \
+            .squeeze(2) \
+            .transpose(0, 1)  # (batch, question_len)
+        alpha = torch.softmax(wg_g, dim=1)  # (batch, question_len)
+        return alpha
+
+
+class UniMatchRNN(torch.nn.Module):
+    r"""
+    interaction context and question with attention mechanism, one direction, using LSTM cell
+    Args:
+        - input_size: The number of expected features in the input Hp and Hq
+        - hidden_size: The number of features in the hidden state Hr
+
+    Inputs:
+        Hp(context_len, batch, input_size): context encoded
+        Hq(question_len, batch, input_size): question encoded
+
+    Outputs:
+        Hr(context_len, batch, hidden_size): question-aware context representation
+        alpha(batch, question_len, context_len): used for visual show
+    """
+
+    def __init__(self, mode, hp_input_size, hq_input_size, hidden_size, gated_attention, enable_layer_norm):
+        super(UniMatchRNN, self).__init__()
+        self.hidden_size = hidden_size
+        self.gated_attention = gated_attention
+        self.enable_layer_norm = enable_layer_norm
+        rnn_in_size = hp_input_size + hq_input_size
+
+        self.attention = MatchRNNAttention(hp_input_size, hq_input_size, hidden_size)
+
+        if self.gated_attention:
+            self.gated_linear = torch.nn.Linear(rnn_in_size, rnn_in_size)
+
+        if self.enable_layer_norm:
+            self.layer_norm = torch.nn.LayerNorm(rnn_in_size)
+
+        self.mode = mode
+        if mode == 'LSTM':
+            self.hidden_cell = torch.nn.LSTMCell(input_size=rnn_in_size, hidden_size=hidden_size)
+        elif mode == 'GRU':
+            self.hidden_cell = torch.nn.GRUCell(input_size=rnn_in_size, hidden_size=hidden_size)
+        else:
+            raise ValueError('Wrong mode select %s, change to LSTM or GRU' % mode)
+        self.reset_parameters()
+
+    def reset_parameters(self):
+        """
+        Here we reproduce Keras default initialization weights to initialize Embeddings/LSTM weights
+        """
+        ih = (param for name, param in self.named_parameters() if 'weight_ih' in name)
+        hh = (param for name, param in self.named_parameters() if 'weight_hh' in name)
+        b = (param for name, param in self.named_parameters() if 'bias' in name)
+        for t in ih:
+            torch.nn.init.xavier_uniform_(t)
+        for t in hh:
+            torch.nn.init.orthogonal_(t)
+        for t in b:
+            torch.nn.init.constant_(t, 0)
+
+    def forward(self, Hp, Hq):
+        batch_size = Hp.shape[1]
+        context_len = Hp.shape[0]
+
+        # init hidden with the same type of input data
+        h_0 = Hq.new_zeros(batch_size, self.hidden_size)
+        hidden = [(h_0, h_0)] if self.mode == 'LSTM' else [h_0]
+        vis_alpha = []
+
+        for t in range(context_len):
+            cur_hp = Hp[t, ...]  # (batch, input_size)
+            attention_input = hidden[t][0] if self.mode == 'LSTM' else hidden[t]
+
+            alpha = self.attention.forward(cur_hp, Hq, attention_input)  # (batch, question_len)
+            vis_alpha.append(alpha)
+
+            question_alpha = torch.bmm(alpha.unsqueeze(1), Hq.transpose(0, 1)) \
+                .squeeze(1)  # (batch, input_size)
+            cur_z = torch.cat([cur_hp, question_alpha], dim=1)  # (batch, rnn_in_size)
+
+            # layer normalization
+            if self.enable_layer_norm:
+                cur_z = self.layer_norm(cur_z)  # (batch, rnn_in_size)
+
+            cur_hidden = self.hidden_cell.forward(cur_z, hidden[t])  # (batch, hidden_size), when lstm output tuple
+            hidden.append(cur_hidden)
+
+        vis_alpha = torch.stack(vis_alpha, dim=2)  # (batch, question_len, context_len)
+
+        hidden_state = list(map(lambda x: x[0], hidden)) if self.mode == 'LSTM' else hidden
+        result = torch.stack(hidden_state[1:], dim=0)  # (context_len, batch, hidden_size)
+        return result, vis_alpha
+
+
+class MatchRNN(torch.nn.Module):
+    r"""
+    interaction context and question with attention mechanism
+    Args:
+        - input_size: The number of expected features in the input Hp and Hq
+        - hidden_size: The number of features in the hidden state Hr
+        - bidirectional: If ``True``, becomes a bidirectional RNN. Default: ``False``
+        - gated_attention: If ``True``, gated attention used, see more on R-NET
+
+    Inputs:
+        Hp(context_len, batch, input_size): context encoded
+        Hq(question_len, batch, input_size): question encoded
+        Hp_mask(batch, context_len): each context valued length without padding values
+        Hq_mask(batch, question_len): each question valued length without padding values
+
+    Outputs:
+        Hr(context_len, batch, hidden_size * num_directions): question-aware context representation
+    """
+
+    def __init__(self, mode, hp_input_size, hq_input_size, hidden_size, bidirectional, gated_attention,
+                 dropout_p, enable_layer_norm):
+        super(MatchRNN, self).__init__()
+        self.bidirectional = bidirectional
+        self.num_directions = 1 if bidirectional else 2
+
+        self.left_match_rnn = UniMatchRNN(mode, hp_input_size, hq_input_size, hidden_size, gated_attention,
+                                          enable_layer_norm)
+        if bidirectional:
+            self.right_match_rnn = UniMatchRNN(mode, hp_input_size, hq_input_size, hidden_size, gated_attention,
+                                               enable_layer_norm)
+
+        self.dropout = torch.nn.Dropout(p=dropout_p)
+
+    def forward(self, Hp, Hp_mask, Hq, Hq_mask):
+        Hp = self.dropout(Hp)
+        Hq = self.dropout(Hq)
+
+        left_hidden, left_para = self.left_match_rnn.forward(Hp, Hq)
+        rtn_hidden = left_hidden
+        rtn_para = {'left': left_para}
+
+        if self.bidirectional:
+            Hp_inv = masked_flip(Hp, Hp_mask, flip_dim=0)
+            right_hidden_inv, right_para_inv = self.right_match_rnn.forward(Hp_inv, Hq)
+
+            # flip back to normal sequence
+            right_alpha_inv = right_para_inv['alpha']
+            right_alpha_inv = right_alpha_inv.transpose(0, 1)  # make sure right flip
+            right_alpha = masked_flip(right_alpha_inv, Hp_mask, flip_dim=2)
+            right_alpha = right_alpha.transpose(0, 1)
+
+            # right_gated_inv = right_para_inv['gated']
+            # right_gated_inv = right_gated_inv.transpose(0, 1)
+            # right_gated = masked_flip(right_gated_inv, Hp_mask, flip_dim=2)
+            # right_gated = right_gated.transpose(0, 1)
+
+            right_hidden = masked_flip(right_hidden_inv, Hp_mask, flip_dim=0)
+
+            rtn_para['right'] = {'alpha': right_alpha} #, 'gated': right_gated}
+            rtn_hidden = torch.cat((left_hidden, right_hidden), dim=2)
+
+        real_rtn_hidden = Hp_mask.transpose(0, 1).unsqueeze(2) * rtn_hidden
+        last_hidden = rtn_hidden[-1, :]
+
+        return real_rtn_hidden, last_hidden, rtn_para
